Clean up unused code and renames
