--- spark-cypher/src/test/scala/org/opencypher/spark/impl/acceptance/MultigraphProjectionBehaviour.scala
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2016-2018 "Neo4j, Inc." [https://neo4j.com]
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.opencypher.spark.impl.acceptance
-
-import org.opencypher.okapi.api.graph.GraphName
-import org.opencypher.okapi.api.value.CypherValue._
-import org.opencypher.spark.impl.CAPSConverters._
-import org.opencypher.spark.impl.CAPSGraph
-
-import scala.collection.immutable.Bag
-
-trait MultigraphProjectionBehaviour { this: AcceptanceTest =>
-
-  def multigraphProjectionBehaviour(initGraph: String => CAPSGraph): Unit = {
-    def testGraph1 = initGraph("CREATE (:Person {name: 'Mats'})")
-    def testGraph2 = initGraph("CREATE (:Person {name: 'Phil'})")
-    def testGraph3 = initGraph("CREATE (:Car {type: 'Toyota'})")
-
-  }
-
-}
